requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-Limiter==2.4.0
Flask-HTTPAuth==4.5.0
Flask-SQLAlchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.0
openai==0.27.0
Werkzeug==2.0.1
docker==5.0.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
